Dev - Implemented Delete Dialog Confirmation

+++ src/app/broker/broker.component.ts
@@ -3,6 +3,9 @@ import { NavigationStart, Router, RouterEvent } from '@angular/router';
 import { Broker } from './Broker';
 import { BrokerService } from './broker.service';
 import { filter } from 'rxjs';
+import { openUpdateBrokerDialog } from './update-broker/update-broker.component';
+import { MatDialog } from '@angular/material/dialog';
+import { openDeleteBrokerDialog } from './delete-broker/delete-broker.component';
 
 @Component({
   selector: 'app-broker',
@@ -11,7 +14,9 @@ import { filter } from 'rxjs';
 })
 export class BrokerComponent implements OnInit{
 
-  constructor(private brokerService: BrokerService, private router: Router) {
+  constructor(private brokerService: BrokerService,
+              private router: Router,
+              private dialog: MatDialog) {
     //Subscribes to router events to update brokers list
     router.events
     .pipe(filter(event => event instanceof NavigationStart))
@@ -38,14 +43,33 @@ export class BrokerComponent implements OnInit{
   }
 
   editBroker(broker: Broker){
-    this.router.navigate(['/update-broker']);
-    this.brokerService.setBrokerToBeUpdated(broker);
+    openUpdateBrokerDialog(this.dialog, broker)
+      .pipe(
+          filter(val => !!val)
+      )
+      .subscribe(
+        brokerForm => {
+          broker.name = brokerForm['name'];
+          broker.description = brokerForm['description'];
+          this.brokerService.updateBroker(broker.id, broker)
+            .subscribe(
+              result => this.getAllBrokers()
+            )
+        }
+      );
   }
 
   deleteBroker(brokerId: number){
-    this.brokerService.deleteBroker(brokerId).subscribe(
-      data => this.getAllBrokers()
-    );
+
+    openDeleteBrokerDialog(this.dialog).subscribe(
+      val => {
+        if(val === true){
+          this.brokerService.deleteBroker(brokerId).subscribe(
+            data => this.getAllBrokers()
+          );
+        }
+      }
+    )
   }
